+++ lib/rails_memory_profiler/notifiers/logger.rb
@@ -1,0 +1,9 @@
+module RailsMemoryProfiler
+  module Notifiers
+    class Logger
+      def call(report)
+        Rails.logger.info(Notifiers.format_line(report))
+      end
+    end
+  end
+end