+++ lib/rails_memory_profiler/notifiers/stdout.rb
@@ -1,0 +1,9 @@
+module RailsMemoryProfiler
+  module Notifiers
+    class Stdout
+      def call(report)
+        $stdout.puts(Notifiers.format_line(report))
+      end
+    end
+  end
+end